:recycle: Refactor NavBar Data
