+++ portal-jtk/src/lib/types/navigation.type.ts
@@ -1,0 +1,10 @@
+export type Navigation = {
+  name: string;
+  links: Array<Link>;
+  url?: string;
+};
+
+type Link = {
+  name: string;
+  url: string;
+};
